Implements an access control list (ACL)
A permission mechanism is added based on the definition of the available
functionalities and their enabled actions. This functionality-action
combination will be associated to a role/group. Every user will have a unique
role associated to it, having the assigned permissions and privileges. To this
end, the classes and resources needed to implement the ACL have been added
including the permission conditions associated with each of the application's
functionalities.

+++ service/src/main/java/org/sing_group/rihana/service/patient/DefaultPatientService.java
@@ -24,12 +24,16 @@ package org.sing_group.rihana.service.patient;
 
 import java.util.stream.Stream;
 
+import javax.annotation.Resource;
 import javax.annotation.security.PermitAll;
+import javax.ejb.EJBAccessException;
+import javax.ejb.SessionContext;
 import javax.ejb.Stateless;
 import javax.inject.Inject;
 
 import org.sing_group.rihana.domain.dao.spi.patient.PatientDAO;
 import org.sing_group.rihana.domain.entities.patient.Patient;
+import org.sing_group.rihana.service.spi.acl.permission.PermissionService;
 import org.sing_group.rihana.service.spi.patient.PatientService;
 
 @Stateless
@@ -39,38 +43,121 @@ public class DefaultPatientService implements PatientService {
 	@Inject
 	private PatientDAO patientDAO;
 
+	@Inject
+	private PermissionService permissionService;
+
+	@Resource
+	private SessionContext context;
+
 	@Override
 	public Patient create(Patient patient) {
+
+		String loginLogged = context.getCallerPrincipal().getName();
+		if (!this.permissionService.hasPermission(
+			loginLogged,
+			"PATIENT_MANAGEMENT",
+			"ADD") &&
+			!this.permissionService.isAdmin(loginLogged)
+		) {
+			throw new EJBAccessException("Insufficient privileges");
+		}
+
 		return patientDAO.create(patient);
 	}
 
 	@Override
 	public Stream<Patient> getPatients() {
+
+		String loginLogged = context.getCallerPrincipal().getName();
+		if (!this.permissionService.hasPermission(
+			loginLogged,
+			"PATIENT_MANAGEMENT",
+			"SHOW_ALL") &&
+			!this.permissionService.isAdmin(loginLogged)
+		) {
+			throw new EJBAccessException("Insufficient privileges");
+		}
+
 		return patientDAO.getPatients();
 	}
 
 	@Override
 	public Patient get(String id) {
+
+		String loginLogged = context.getCallerPrincipal().getName();
+		if (!this.permissionService.hasPermission(
+			loginLogged,
+			"PATIENT_MANAGEMENT",
+			"SHOW_CURRENT") &&
+			!this.permissionService.isAdmin(loginLogged)
+		) {
+			throw new EJBAccessException("Insufficient privileges");
+		}
+
 		return patientDAO.get(id);
 	}
 
 	@Override
 	public Patient getPatientBy(String patientID) {
+
+		String loginLogged = context.getCallerPrincipal().getName();
+		if (!this.permissionService.hasPermission(
+				loginLogged,
+				"PATIENT_MANAGEMENT",
+				"SHOW_CURRENT") &&
+			!this.permissionService.isAdmin(loginLogged)
+		) {
+			throw new EJBAccessException("Insufficient privileges");
+		}
+
 		return patientDAO.getPatientBy(patientID);
 	}
 
 	@Override
 	public boolean existsPatientBy(String patientID) {
+
+		String loginLogged = context.getCallerPrincipal().getName();
+		if (!this.permissionService.hasPermission(
+				loginLogged,
+				"PATIENT_MANAGEMENT",
+				"SHOW_CURRENT") &&
+			!this.permissionService.isAdmin(loginLogged)
+		) {
+			throw new EJBAccessException("Insufficient privileges");
+		}
+
 		return patientDAO.existsPatientBy(patientID);
 	}
 
 	@Override
 	public Patient edit(Patient patient) {
+
+		String loginLogged = context.getCallerPrincipal().getName();
+		if (!this.permissionService.hasPermission(
+				loginLogged,
+				"PATIENT_MANAGEMENT",
+				"EDIT") &&
+			!this.permissionService.isAdmin(loginLogged)
+		) {
+			throw new EJBAccessException("Insufficient privileges");
+		}
+
 		return patientDAO.edit(patient);
 	}
 
 	@Override
 	public void delete(Patient patient) {
+
+		String loginLogged = context.getCallerPrincipal().getName();
+		if (!this.permissionService.hasPermission(
+				loginLogged,
+				"PATIENT_MANAGEMENT",
+				"DELETE") &&
+			!this.permissionService.isAdmin(loginLogged)
+		) {
+			throw new EJBAccessException("Insufficient privileges");
+		}
+
 		patientDAO.delete(patient);
 	}
 }
